Support unique UDP listening port and fix incroment
A new option is introduced to not increase the listening port. This has sense when the received RTP stream is forwarded over NAT or tunneled.
Futhermore this change fixes the trouble created when the change of the UDP port raises the maximum port value. And it adds a safe upper value (MAXLISTPORT) for the number of the concurrent listening RTP connections).